first iteration matches hand calculations. Using the governing equations from google drive, N_pins = 15
